+++ src/Contracts/LineItemType.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Vanilo\Contracts;
+
+interface LineItemType extends \Konekt\Enum\EnumInterface
+{
+    public function isShippable(): bool;
+
+    public function isPhysical(): bool;
+
+    public function isAService(): bool;
+
+    public function isDigital(): bool;
+
+    public function hasPrice(): bool;
+}