+++ src/Contracts/Changelog.md
@@ -7,6 +7,7 @@
 
 - Dropped PHP 8.3 support
 - Dropped Laravel 10 & 11 support
+- Added the `LineItemType` interface
 - Added the konekt/enum package dependency (v4.2)
 
 ## 5.x Series